Restructured packages
core: new package;
Fuzz: new inchoate class.
