+++ src/main/scala/com/phasmidsoftware/number/misc/ContinuedFraction.scala
@@ -1,4 +1,6 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
+
+import com.phasmidsoftware.number.core.Rational
 
 import scala.annotation.tailrec
 import scala.util.Try
