+++ src/main/scala/com/phasmidsoftware/number/misc/DiFunc.scala
@@ -1,4 +1,4 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
 
 /**
   * @author scalaprof
