+++ src/main/scala/com/phasmidsoftware/number/misc/InfiniteSum.scala
@@ -1,4 +1,4 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
 
 /**
   * This case class represents an infinite sum of terms which can be derived by a function (f) which
