+++ src/main/scala/com/phasmidsoftware/number/misc/LazyFunction.scala
@@ -1,4 +1,6 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
+
+import com.phasmidsoftware.number.core.Rational
 
 /**
   * This abstract class is the base class for various functions used by LazyNumber algebra.
