+++ src/main/scala/com/phasmidsoftware/number/misc/LazyNumber.scala
@@ -1,4 +1,6 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
+
+import com.phasmidsoftware.number.core.Rational
 
 /**
   * LazyNumber is an abstract (base) class for Valuable objects.
