+++ src/main/scala/com/phasmidsoftware/number/parse/RationalParser.scala
@@ -1,6 +1,6 @@
 package com.phasmidsoftware.number.parse
 
-import com.phasmidsoftware.number.model.Rational
+import com.phasmidsoftware.number.core.Rational
 
 import scala.util.Try
 import scala.util.parsing.combinator.JavaTokenParsers
