+++ src/test/scala/com/phasmidsoftware/number/misc/LazyNumberFuzzySpec.scala
@@ -1,4 +1,4 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
 
 import org.scalatest.flatspec
 import org.scalatest.matchers.should
