+++ src/test/scala/com/phasmidsoftware/number/misc/LazyNumberRationalSpec.scala
@@ -1,5 +1,6 @@
-package com.phasmidsoftware.number.model
+package com.phasmidsoftware.number.misc
 
+import com.phasmidsoftware.number.core.Rational
 import org.scalatest.flatspec
 import org.scalatest.matchers.should
