+++ src/test/scala/com/phasmidsoftware/number/parse/FuzzyParserSpec.scala
@@ -1,6 +1,6 @@
 package com.phasmidsoftware.number.parse
 
-import com.phasmidsoftware.number.model.{Exact, Gaussian}
+import com.phasmidsoftware.number.misc.{Exact, Gaussian}
 import org.scalatest.flatspec
 import org.scalatest.matchers.should
