+++ src/test/scala/com/phasmidsoftware/number/parse/RationalParserSpec.scala
@@ -1,7 +1,6 @@
 package com.phasmidsoftware.number.parse
 
-import com.phasmidsoftware.number.model.Rational
-import com.phasmidsoftware.number.parse.RationalParser
+import com.phasmidsoftware.number.core.Rational
 import org.scalatest.flatspec
 import org.scalatest.matchers.should
